Better waiting for datanode proxy calls
