+++ full-backend-tests/src/test/java/org/graylog/datanode/DatanodeOpensearchProxyIT.java
@@ -16,33 +16,23 @@
  */
 package org.graylog.datanode;
 
-import com.github.rholder.retry.Attempt;
 import com.github.rholder.retry.RetryException;
-import com.github.rholder.retry.RetryListener;
-import com.github.rholder.retry.RetryerBuilder;
-import com.github.rholder.retry.StopStrategies;
-import com.github.rholder.retry.WaitStrategies;
 import io.restassured.response.ValidatableResponse;
 import org.assertj.core.api.Assertions;
 import org.graylog.testing.completebackend.Lifecycle;
-import org.graylog.testing.completebackend.apis.GraylogApiResponse;
 import org.graylog.testing.completebackend.apis.GraylogApis;
 import org.graylog.testing.containermatrix.SearchServer;
 import org.graylog.testing.containermatrix.annotations.ContainerMatrixTest;
 import org.graylog.testing.containermatrix.annotations.ContainerMatrixTestsConfiguration;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 
 @ContainerMatrixTestsConfiguration(serverLifecycle = Lifecycle.CLASS, searchVersions = SearchServer.DATANODE_DEV, additionalConfigurationParameters = {@ContainerMatrixTestsConfiguration.ConfigurationParameter(key = "GRAYLOG_DATANODE_PROXY_API_ALLOWLIST", value = "true")})
 public class DatanodeOpensearchProxyIT {
 
-    private static final Logger LOG = LoggerFactory.getLogger(DatanodeOpensearchProxyIT.class);
     private GraylogApis apis;
 
     @BeforeEach
@@ -52,7 +42,7 @@ public class DatanodeOpensearchProxyIT {
 
     @ContainerMatrixTest
     void testProxyPlaintextGet() throws ExecutionException, RetryException {
-        waitForLeader();
+        apis.datanodeProxy().waitForLeader();
         final ValidatableResponse response = apis.get("/datanodes/leader/request/_cat/indices", 200);
         final String responseBody = response.extract().body().asString();
         Assertions.assertThat(responseBody).contains(".ds-gl-datanode-metrics").contains("graylog_0").contains("gl-system-events_0");
@@ -72,41 +62,10 @@ public class DatanodeOpensearchProxyIT {
 
     @ContainerMatrixTest
     void testTargetSpecificDatanodeInstance() {
-        final List<String> datanodes = getDatanodes().properJSONPath().read("elements.*.hostname");
+        final List<String> datanodes = apis.datanodeProxy().getDatanodes().properJSONPath().read("elements.*.hostname");
         Assertions.assertThat(datanodes).isNotEmpty();
 
         final String hostname = datanodes.iterator().next();
         apis.get("/datanodes/" + hostname + "/request/_mapping", 200).assertThat().body("graylog_0.mappings.properties.gl2_accounted_message_size.type", Matchers.equalTo("long"));
-    }
-
-    private GraylogApiResponse getDatanodes() {
-        return new GraylogApiResponse(apis.get("/system/cluster/datanodes", 200));
-    }
-
-    /**
-     * A datanode becomes leader if the underlying opensearch node becomes leader. This may take a while during the
-     * startup. Our periodical checks the leader status every 10s, so in this interval there should appear one leader
-     * for sure.
-     *
-     * @see org.graylog.datanode.periodicals.ClusterManagerDiscovery
-     */
-    private void waitForLeader() throws ExecutionException, RetryException {
-        RetryerBuilder.<GraylogApiResponse>newBuilder()
-                .retryIfResult(r -> !isAnyLeader(r))
-                .withRetryListener(new RetryListener() {
-                    @Override
-                    public <V> void onRetry(Attempt<V> attempt) {
-                        LOG.info("Waiting for one datanode to become leader, attempt {}", attempt.getAttemptNumber());
-                    }
-                })
-                .withWaitStrategy(WaitStrategies.fixedWait(1, TimeUnit.SECONDS))
-                .withStopStrategy(StopStrategies.stopAfterAttempt(10))
-                .build()
-                .call(this::getDatanodes);
-    }
-
-    private boolean isAnyLeader(GraylogApiResponse r) {
-        List<Boolean> isLeader = r.properJSONPath().read("elements.*.is_leader");
-        return isLeader.stream().anyMatch(leader -> leader);
     }
 }
